ai-tests/requirements.txt
playwright==1.48.0
openai==1.54.0
python-dotenv==1.0.0
faker==30.8.2

